Add files via upload
I have taken two PCs and connected them with a switch. I configured the IP address of each pc and then tried to connect with the other pc using the ping command in the terminal. 4 packets will be sent to the other pc and they will be received back indicating successful transfer of data.